Defrag the read buffer when it gets full.
If we are reading many packets of different sizes from a fast database
server, it's possible that the skip() method of the packet reader will
never find the start and end pointers aligned and so it will never reset
them. This can lead to throwing 'Socket receive buffer full' errors when
we reach the end, even if there's plenty of unused space at the start of
the buffer from previous packets.

This checks for this situation and moves the live data and current
pointers to the start of the buffer, thus making room (if possible).